openai server trait and move HTTP endpoints to `crates/runtime/src/http/v1/` (#1619)
* openai server trait and move HTTP endpoints to crates/runtime/src/http/v1/

* fixes in for problem with deserialisation and model swap out

* clippy and lint
