+++ crates/runtime/src/http/routes.rs
@@ -14,8 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+use crate::model::LLMModelStore;
 use crate::{config, datafusion::DataFusion};
-use crate::{EmbeddingModelStore, LLMModelStore};
+use crate::{EmbeddingModelStore, OpenaiServerStore};
 use app::App;
 use axum::routing::patch;
 use model_components::model::Model;
@@ -35,12 +36,14 @@ use tokio::{sync::RwLock, time::Instant};
 
 use super::v1;
 
+#[allow(clippy::too_many_arguments)]
 pub(crate) fn routes(
     app: Arc<RwLock<Option<App>>>,
     df: Arc<DataFusion>,
     models: Arc<RwLock<HashMap<String, Model>>>,
     llms: Arc<RwLock<LLMModelStore>>,
     embeddings: Arc<RwLock<EmbeddingModelStore>>,
+    openai_servers: Arc<RwLock<OpenaiServerStore>>,
     config: Arc<config::Config>,
     with_metrics: Option<SocketAddr>,
 ) -> Router {
@@ -66,10 +69,12 @@ pub(crate) fn routes(
             .route("/v1/models/:name/predict", get(v1::inference::get))
             .route("/v1/predict", post(v1::inference::post))
             .route("/v1/nsql", post(v1::nsql::post))
-            .route("/v1/embed", post(v1::embed::post))
+            .route("/v1/chat/completions", post(v1::chat::post))
+            .route("/v1/embeddings", post(v1::embeddings::post))
             .route("/v1/assist", post(v1::assist::post))
             .layer(Extension(llms))
             .layer(Extension(models))
+            .layer(Extension(openai_servers))
             .layer(Extension(embeddings));
     }
